Cancel regular orders without applications and increase head hunt order price

+++ backend/src/main/java/ru/ifmo/se/johnwick/model/dto/HeadHuntOrderDto.java
@@ -13,6 +13,6 @@ import lombok.experimental.SuperBuilder;
 @EqualsAndHashCode(callSuper = true)
 public class HeadHuntOrderDto extends OrderDto {
     private UserDto succeededKiller;
-    private Double currentPrice;
+    private long currentPrice;
     private String customerName;
 }
